{{cookiecutter.repo_name}}/requirements.txt
muffin==0.0.105
muffin-jinja2==0.0.11
muffin-peewee==0.0.46
muffin-session==0.0.8
